+++ CHANGELOG.md
@@ -9,6 +9,16 @@ the client will treat as a sign to stop retrying. The client will then
 disconnect from the room, instead of remaining in `"connecting"` status
 indefinitely.
 
+### `@liveblocks/react`
+
+Fix a bug with `useSelf()` where it would not correctly re-render after entering
+an empty room. It’s now consistent again with `useMyPresence()`.
+
+### DevTools
+
+Fix a bug in the Liveblocks [DevTools](https://liveblocks.io/devtools) panel
+where the "me" view would incorrectly stay empty after entering an empty room.
+
 # v1.1.6
 
 ### `@liveblocks/*`